// src/user/dto/update-user.dto.ts
import { PartialType } from '@nestjs/mapped-types';
import { CreateUserDto } from './create-user.dto';

export class UpdateUserDto extends PartialType(CreateUserDto) {
  name?: string | undefined;
  email?: string | undefined;
  password?: string | undefined;
  currentPassword?: string | undefined;
}